requirements.txt
pandas
numpy
scikit-learn
requests
python-dateutil
joblib
fastparquet
